+++ Apps/AppManager/Source/Malterlib_Cloud_App_AppManagerApp.h
@@ -22,7 +22,7 @@ namespace NMib::NCloud
 		NStr::CStr m_ContainerName;
 		NStr::CStr m_Image;
 		NStr::CStr m_Network;
-		uint32 m_MemoryLimitMB = 0;
+		uint64 m_MemoryLimitMB = 0;
 		fp64 m_CPULimit = 0.0;
 		bool m_bReadOnly = false; /// Read-only container filesystem; all writes must go to mounted paths
 		NContainer::TCMap<NStr::CStr, NStr::CStr> m_Mounts; /// Host path to path inside the container
